structure => relocate files
