+++ src/api/role/entities/role.entity.ts
