+++ src/api/role/role.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { RoleService } from './role.service';
-import { RoleController } from './role.controller';
+import { RoleService } from './service/role.service';
+import { RoleController } from './controller/role.controller';
 import { PrismaModule } from 'src/prisma/prisma.module';
 
 @Module({
